+++ assets/Components/TableData/TableData.jsx
@@ -323,7 +323,7 @@ const ColumnSubMenu = ({ cellOptions, sxButton, onMenuItemClick, columnsPinnedTo
   return (
     <>
       <Tooltip title="Column options">
-        <IconButton sx={sxButton} size="small" onClick={handleOpenMenu}>
+        <IconButton sx={sxButton} size="small" onClick={handleOpenMenu} color="inherit">
           <MoreVertIcon fontSize="small" color="black" />
         </IconButton>
       </Tooltip>
@@ -566,7 +566,7 @@ const EnhancedTableHead = ({
                 onMouseLeave={() => setShowColumnSubmenu(null)}
               >
                 <Box
-                  className={`${commonStyles.flex} ${commonStyles.alignCenter} 
+                  className={`${commonStyles.flex} ${commonStyles.alignCenter}
                   ${
                     headCell.alignContent
                       ? (headCell.alignContent === 'center' && commonStyles.justifyCenter) ||
@@ -798,6 +798,10 @@ export default function TableData({
   const rowsPerPagePaginationOptions = [rowsPerPage, 20, 50, 100].filter((el, idx, self) => self.indexOf(el) === idx).sort((a, b) => a - b);
   const rowRef = useRef(null);
   const sortableColumns = headCells.filter((cell) => cell.sortable).map((cell) => cell.id);
+
+  useEffect(() => {
+    setPage(Number(paginationParams.page - 1));
+  }, [paginationParams.page]);
 
   useEffect(() => {
     if (isNaN(page)) {