+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Custom logos support in dark mode
+- Setting placeholders for creatives
 ### Changed
 - AdServers' name on flow chart label
 ### Fixed